Add files via upload
In this Assignment, I have tried to fit Naïve Bayes model on Donors Chose Dataset.
Donors Chose is a US Based NGO which brings together Teachers and Donors where teachers may request donations for the projects of their students from the donors and there are volunteers who decide which projects must be shown on the website by manually reviewing them. Since the volume of the projects have increased a lot over the past few years, they are trying to automate the review process to reduce load on the volunteers. 

In the project. I have used Bag Of Words and TFIDF to vectorize the text features and played with different n-grams to and dimensionalities to check whether they have any impact upon the AUC Score (Metric). 
More details mentioned in the project itself.